+++ db/migrate/20200414101829_add_previous_identifier_to_activities.rb
@@ -1,0 +1,5 @@
+class AddPreviousIdentifierToActivities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :activities, :previous_identifier, :string
+  end
+end