+++ CHANGELOG.md
@@ -103,6 +103,9 @@
 
 ## [unreleased]
 
+- Activities that have an identifier already can use it as the iati-identifier
+  in the xml output
+
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-4...HEAD
 [release-4]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-3...release-4
 [release-3]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-2...release-3